feat: add abiltity to deserialize tx into json
